overflow fix

+++ libft/ft_atoi.c
@@ -6,14 +6,14 @@
 /*   By: hkovac <hkovac@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/09/08 10:13:30 by hkovac            #+#    #+#             */
-/*   Updated: 2022/01/06 13:55:05 by hkovac           ###   ########.fr       */
+/*   Updated: 2022/01/10 11:20:58 by hkovac           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-long int	ft_atoi(const char *nptr)
+long long	ft_atoi(const char *nptr)
 {
-	long	result;
-	int		signe;
+	long long	result;
+	long long	signe;
 
 	result = 0;
 	signe = 1;
@@ -27,7 +27,7 @@ long int	ft_atoi(const char *nptr)
 	}
 	while (*nptr >= '0' && *nptr <= '9')
 		result = result * 10 + (*nptr++ - '0');
-	return (signe * result);
+	return ((long long int)signe * result);
 }
 /*
 #include <stdio.h>
